I edited the FrmMemberData.cs on btnSave_click method on ScoreInfo on Average because when you click the save button the average in the database is always 0.
I edited the CalculateHandicapPins on the Calculation.cs , so it will not return a negative Handicap.

I added the CalculateHandicapPins method on the MemberDb.cs on AddMember method, so when a new member signs up for their 1st tournament, a handicap is assigned from the league average that is entered into the program by the user. As long as the new member have not participated in a tournament(or have a 0 tournament average), the handicap can be updated if the league average is updated.

I  added a comment in FinalizeTempDB.cs on AddFinalizeTemp. The comment is the code to update the handicap of a member which is not working. I believe it is a problem in the database.