Refactor Tournament Results (#1119)
* Remove the incomplete and unused invalid member flow in the MemberImport program

* Remove unused TextChanged event for tbClientInputCount

The `TextChanged` event handler for the `tbClientInputCount` TextBox was removed from `FrmTournamentResults.Designer.cs`. This change eliminates unnecessary functionality that was previously triggered on text changes. A comment in `FrmTournamentResults.cs` indicating that the method was created by accident was also removed. The `KeyDown` event handler for the TextBox remains intact, preserving the functionality for handling key presses.

* Fix earnings formatting in Excel output

Enhanced the formatting of the earnings value in the Excel sheet.
Replaced the direct assignment of the earnings value with a new block of code that checks for null and attempts to parse the value as a double.
If parsing is successful, the value is formatted as currency with zero decimal places; otherwise, it falls back to the original string or defaults to "$0" if null.
This change improves the presentation of earnings in the output.

* Fix handicap and Total Score outputs to correct columns

* Add placestanding output to right side

* Progressive pot fixed, new issue with finish placing

* Add row handling for Excel output in tournament results

Implement a loop to manage DataTable rows in the Excel worksheet.
Insert new rows and merge columns G and H when the row count exceeds 35,
ensuring organized output for cases with many results.

* Refactor loop structure and row handling in results

Changed from a `for` loop to a `while` loop for better control over iteration. Updated row calculations to consistently use `startRow` instead of `excelRow`. Adjusted logic for inserting rows and merging cells to ensure correct manipulation of rows. Modified handling of progressive pot earnings to display the correct participant's data without advancing the index incorrectly. Added additional cell assignments for `startRow` to capture all relevant data.

* Refactor progressive pot output handling

Updated logic to check for progressive pot rows by evaluating the next row instead of the current one. Removed old checks for specific `startRow` values and implemented a new approach to retrieve and display the side pot value for the current bowler.

* Rename startRow to excelRow and update logic

Refactor the method to use `excelRow` for Excel operations instead of `startRow`. Update row insertion and merging logic accordingly. Adjust value handling for various columns to ensure correct data assignment. Modify payout formula to include `sidePotValue` for accurate total calculations.

* Add ordinal formatting with tie indicator in export

Implemented `GetOrdinalWithTie` method for converting numeric positions to ordinal strings, including tie indicators. Updated `ExportToExcel` to use this method for formatting place standings, incorporating logic to check for ties based on adjacent rows.